Customer Class Created

+++ C0751731_MID_TERM_MAD3004/C0751731_MID_TERM_MAD3004/Bill.swift
@@ -13,9 +13,9 @@ class Bill : IDisplay
     var bId : Int?
     var bDate : Date?
     var bType : String
-    var bTotalAmount : Double?
+    var bTotalAmount : Float?
     
-    init(bId : Int,bDate : Date,bType : String, bTotalAmount : Double)
+    init(bId : Int,bDate : Date,bType : String, bTotalAmount : Float)
     {
         self.bId = 0
         self.bDate = bDate
